[blobDescriptor]: add length descriptor for affordance experiments
For now the length of a blob is defined as the major axis of the enclosing rectangle. Depending on the tests we might change the definition.

The vector of descriptors is now: (area convexity eccentricity compactness circleness squareness length).

Note: the previous "elongatedness" is only used internally for visualization purposes, it is not present in the vector of descriptors anymore (replaced by lenght).

/cc @lorejam @atabakd