[world state]: improve debug prints of interaction with activeParticleTrack
Previously, worldStateManager was sending "pause" and "resume" RPC commands
to activeParticleTrack and was simply accepting the response "ok".

With this commit more information is printed, depending if the response in
the RPC is "ok", or "fail", or otherwise.

Related to #181 and https://github.com/robotology/poeticon/commit/b921e3768a83b9086c7fe45a7a53a43fdd53478d