[world state]: make "init" return immediately when already initialized
When receiving subsequence "init" commands (after initialization has already
been done), worldStateManager was attempting lengthy initialization procedures.

This return statement prevents that and speeds this case up.

For now, the returned value in this case is true/"ok". We can change it to
false/"fail" (conceptually more correct), provided that we change the planner
side as well, PlannerThread::startPlanning().