[world state]: improve update routine
Previously, if an object was updated while it was invisible (not tracked), some symbols were wrongly set to zero (pos2d, desc2d, tooldesc2d). This commit corrects that, by keeping the current value existing in the short-term memory for those fields in this situation. This ensures that sensible coordinates and shape descriptors are used in the "askForTool" instruction and in affordance network queries, respectively.