[world state]: distinguish between geometric and semantic changes
Geometric change: object not visible and occluded/stacked.

Semantic change: object not visible and not in any stack -> really gone from scene.

In this commit the above situations are determined and the information is printed, but not yet propagated to the rest of the architecture (will do that in another commit after tests).