[world state]: fix hand is_free value after reset
Previously, the value of the hand is_free boolean symbol after a world state "reset" was erroneously set to true regardless of the current hand state, and then a world state "update" would set the correct value.

This commit prevents that situation: now during a "reset" the is_free value is immediately set to the current hand status (obtained by activityInterface) without needing the extra "update".

This was not a serious problem in the first place, because the planner is always sending first a "reset" then an "update" to the world state, but it's more consistent now.

/cc @vtikha @lorejam