planner: fix drop action RPC request
This commit ensures that when the planner requests the drop action to activityInterface via RPC, it does that with the required syntax: "drop objName".

Before this fix, a wrong RPC syntax was used containing a spurious third argument "handName", inherited from the previous action. This led to strange behaviour on the activityInterface side: the drop action itself worked, but the subsequent action (e.g., take) would get stuck.

(Also, this commit clears the bottle used for storing the RPC reply received by opc2prada, for safety.)

/cc @AlexAntn